Reorganization of component and service folders In Project
