+++ src/app/pages/dashboard/dashboard.component.spec.ts
@@ -5,7 +5,7 @@ import { NgChartsModule } from 'ng2-charts';
 import { BarChartComponent } from 'src/app/shared/components/bar-chart/bar-chart.component';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { DashboardRoutingModule } from './dashboard-routing.module';
-import { MenuService } from 'src/app/shared/services/menu.service';
+import { MenuService } from 'src/app/shared/services/menu-service/menu.service';
 
 describe('DashboardComponent', () => {
   let component: DashboardComponent;
