+++ src/app/pages/despesas/despesas.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { DespesasComponent } from './despesas.component';
 import { CommonModule } from '@angular/common';
 import { SharedModule } from 'src/app/shared/shared.module';
-import { MenuService } from 'src/app/shared/services/menu.service';
+import { MenuService } from 'src/app/shared/services/menu-service/menu.service';
 import { RouterTestingModule } from '@angular/router/testing';
 import { DespesasRoutingModule } from './despesas-routing.module';
