+++ src/app/shared/shared.module.ts
@@ -1,8 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LayoutComponent } from './components/layout/layout.component';
-
-
 @NgModule({
   declarations: [LayoutComponent],
   exports: [LayoutComponent],
